backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
pydantic==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
litellm==1.50.0
pypdf2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
Pillow==10.4.0
cryptography==43.0.1
aiofiles==24.1.0
httpx==0.27.2
httpx-sse>=0.4.0
mcp>=1.0.0
duckduckgo-search==6.3.7
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
lark-oapi>=1.3.0
